Fix dir creation

+++ install-dotfiles.sh
@@ -6,8 +6,8 @@ HOMEDIR="$REAL_HOMEDIR/home"
 DOTFILES_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 
 SYMLINKS_DIRS=( 
-    "at_login.sh"
-    "awesome"
+	"at_login.sh"
+	"awesome"
 	"nvim"
 	"zsh"
 	"alacritty"
@@ -18,49 +18,49 @@ SYMLINKS_DIRS=(
 	"gtk-3.0"
 	"gtk-4.0"
 	"redshift"
-    "copyq"
-    "keepassxc"
-    "fish"
-    "xsessions"
-    "cmus/autosave"
-    "cmus/red_theme.theme"
-    "cmus/notify.sh"
-    "topgrade.toml"
-    "neofetch"
-    "chromium/Default/Extensions"
+	"copyq"
+	"keepassxc"
+	"fish"
+	"xsessions"
+	"cmus/autosave"
+	"cmus/red_theme.theme"
+	"cmus/notify.sh"
+	"topgrade.toml"
+	"neofetch"
+	"chromium/Default/Extensions"
 	"chromium/Default/Extension State"
-    "chromium/Default/Sync Extension Settings"
-    "chromium/Default/Managed Extension Settings"
-    "chromium/Default/Local Extension Settings"
-    "BetterDiscord/plugins"
+	"chromium/Default/Sync Extension Settings"
+	"chromium/Default/Managed Extension Settings"
+	"chromium/Default/Local Extension Settings"
+	"BetterDiscord/plugins"
 )
 
 REAL_HOME_DIRS=(
-    ".bashrc"
+	".bashrc"
 )
 # If path starts with %, will not override
 COPY_DIRS=(
-    "chromium/Default/Preferences"
-    "%chromium/Default/Cookies"
+	#"chromium/Default/Preferences"
+	#"%chromium/Default/Cookies"
     "chromium/Local State"
-    "tutanota-desktop/conf.json"
-    "discord/settings.json"
-    "FreeTube/settings.db"
+	"tutanota-desktop/conf.json"
+	"discord/settings.json"
+	"FreeTube/settings.db"
 )
 
 
 LINK_HOME_DIRS=(
-    ".config"
-    ".local"
-    "Documents"
-    "Downloads"
-    "Pictures"
-    "Videos"
-    "Programming"
-    "VM"
-    "Games"
-    "Temp"
-    "Music"
+	".config"
+	".local"
+	"Documents"
+	"Downloads"
+	"Pictures"
+	"Videos"
+	"Programming"
+	"VM"
+	"Games"
+	"Temp"
+	"Music"
 )
 
 LGREEN='\033[1;32m'
@@ -69,38 +69,38 @@ LBLUE='\033[1;34m'
 NC='\033[0m' 
 
 function confirm() {
-    echo -en "$LBLUE |||$GREEN Do you want to override ${LGREEN}$1 $2 $3 $LBLUE(Y/n)? >> $NC"
-    if [ ! -z $YOLO ] && [ $YOLO -eq 1 ]; then
-        echo "y"
-        rm -rf $1 $2 $3
-        return
-    fi
-    read choice
-    case "$choice" in 
-    y|Y|"" ) rm -rf $1 $2 $3;;
-    n|N ) return;;
-    * ) confirm $1 $2 $3; return;;
-    esac
+	echo -en "$LBLUE |||$GREEN Do you want to override ${LGREEN}$1 $2 $3 $LBLUE(Y/n)? >> $NC"
+	if [ ! -z $YOLO ] && [ $YOLO -eq 1 ]; then
+		echo "y"
+		rm -rf $1 $2 $3
+		return
+	fi
+	read choice
+	case "$choice" in 
+	y|Y|"" ) rm -rf $1 $2 $3;;
+	n|N ) return;;
+	* ) confirm $1 $2 $3; return;;
+	esac
 }
 
 mkdir -p $REAL_HOMEDIR
 mkdir -p $HOMEDIR
 
 for dir in "${LINK_HOME_DIRS[@]}"; do
-    DEST="$REAL_HOMEDIR/$dir"
-    if [ -h "$DEST" ]; then unlink "$DEST"; fi
-    if [ -e "$DEST" ]; then confirm $DEST; fi
-    mkdir -p $HOMEDIR/$dir
-    ln -sfT $HOMEDIR/$dir $DEST
+	DEST="$REAL_HOMEDIR/$dir"
+	if [ -h "$DEST" ]; then unlink "$DEST"; fi
+	if [ -e "$DEST" ]; then confirm $DEST; fi
+	mkdir -p $HOMEDIR/$dir
+	ln -sfT $HOMEDIR/$dir $DEST
 done
 
 
 for dir in "${REAL_HOME_DIRS[@]}"; do
-    FROM="$DOTFILES_DIR/dotfiles/$dir"
-    DEST="$REAL_HOMEDIR/$dir"
-    if [ -h "$DEST" ]; then unlink "$DEST"; fi
-    if [ -e "$DEST" ]; then confirm $DEST; fi
-    mkdir -p "$(dirname $DEST)"
+	FROM="$DOTFILES_DIR/dotfiles/$dir"
+	DEST="$REAL_HOMEDIR/$dir"
+	if [ -h "$DEST" ]; then unlink "$DEST"; fi
+	if [ -e "$DEST" ]; then confirm $DEST; fi
+	mkdir -p "$(dirname $DEST | head --lines 1)"
 	ln -sfT "$FROM" "$DEST"
 	chown -R $USER1:$USER1 "$DEST"
 done
@@ -108,30 +108,31 @@ done
 for dir in "${SYMLINKS_DIRS[@]}"; do
 	FROM="$DOTFILES_DIR/dotfiles/$dir"
 	DEST="$HOMEDIR/.config/$dir"
-    if [ -h "$DEST" ]; then unlink "$DEST"; fi
-    if [ -e "$DEST" ]; then confirm $DEST; fi
-    mkdir -p "$(dirname $DEST)"
+	if [ -h "$DEST" ]; then unlink "$DEST"; fi
+	if [ -e "$DEST" ]; then confirm $DEST; fi
+	mkdir -p "$(dirname $DEST | head --lines 1)"
 	ln -sfT "$FROM" "$DEST"
 	chown -R $USER1:$USER1 "$DEST"
 done
 
-for dir in "${COPY_DIRS[@]}"; do
-    if [[ $dir = %* ]]; then
-        dir="${dir:1}"
-        FROM="$DOTFILES_DIR/dotfiles/$dir"
-        DEST="$HOMEDIR/.config/$dir"
-        if [ ! -e "$DEST" ]; then
-            cp -rf "$FROM" "$DEST"
-        fi
-        continue
-    fi
-	FROM="$DOTFILES_DIR/dotfiles/$dir"
-	DEST="$HOMEDIR/.config/$dir"
-    if [ -h "$DEST" ]; then unlink "$DEST"; fi
-    if [ -e "$DEST" ]; then confirm $DEST; fi
-    mkdir -p "$(dirname $DEST)"
-	cp -rf "$FROM" "$DEST"
-	chown -R $USER1:$USER1 "$DEST"
+for DIR in "${COPY_DIRS[@]}"; do
+	if [[ $DIR = %* ]]; then
+		DIR="${DIR:1}"
+		echo $DIR
+		FROM="$DOTFILES_DIR/dotfiles/$DIR"
+		DEST="$HOMEDIR/.config/$DIR"
+		if [ ! -e "$DEST" ]; then
+			cp -rf "$FROM" "$DEST"
+		fi
+	else
+		FROM="$DOTFILES_DIR/dotfiles/$DIR"
+		DEST="$HOMEDIR/.config/$DIR"
+		if [ -h "$DEST" ]; then unlink "$DEST"; fi
+		if [ -e "$DEST" ]; then confirm $DEST; fi
+		mkdir -p "$(dirname $DEST | head --lines 1)"
+		cp -rf "$FROM" "$DEST"
+		chown -R $USER1:$USER1 "$DEST"
+	fi
 done
 
 # Update nvim plugins
